Ref: #9 Added moneybag image to profile instead of coins and exp to the exp amount

+++ src/main/java/io/github/mannjamin/ducky/database/manager/data/RankUp.java
@@ -2,6 +2,6 @@ package io.github.mannjamin.ducky.database.manager.data;
 
 public class RankUp {
     public Integer rank;
-    public Long expRequired;
+    public Integer expRequired;
     public String rankName;
 }
